add 'dummy' architecture
'dummy' architecture is a way to build koowaldah kernel into a userspace
executable and be run in another operating system (currently, linux)
environment.

The kernel itself is linked into a shared object file; it gets dlopen'ed
by 'kos-loader' which takes care to provide the kernel with all the
necessary calls and stuff.

Signed-off-by: Alexander Shishkin <alexander.shishckin@gmail.com>