i386: add image loader
This adds a so-called loader code that is prefixed to kos kernel binary
and implements x86 linux boot protocol. In the end of build process,
a kos-image is produced (analogous to zImage in linux) that can be loaded
into qemu with -kernel option, that is, without the need for grub (or
any other) bootloader. A number of bootloaders can load this image as
well (syslinux, lilo, bootsector).

Signed-off-by: Alexander Shishkin <ash@koowaldah.org>